convert Zomcav into maven module
