+++ Zomcav/pom.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>com.let.us.java</groupId>
+    <artifactId>LetUsJava</artifactId>
+    <version>1.0-SNAPSHOT</version>
+  </parent>
+
+  <groupId>com.let.us.java</groupId>
+  <artifactId>Zomcav</artifactId>
+  <packaging>jar</packaging>
+
+  <name>Zomcav(Codechef)</name>
+
+</project>
